+++ asdf_astropy/io/tests/test_io.py
@@ -14,33 +14,33 @@ def make_table():
     return Table([a, b, c], names=("a", "b", "c"), meta={"name": "first table"})
 
 
-def test_table_io(tmpdir):
-    tmpfile = str(tmpdir.join("table.asdf"))
+def test_table_io(tmp_path):
+    tmp_file = tmp_path / "table.asdf"
 
     table = make_table()
 
-    table.write(tmpfile)
+    table.write(tmp_file)
 
     # Simple sanity check using ASDF directly
-    with asdf.open(tmpfile) as af:
+    with asdf.open(tmp_file) as af:
         assert "data" in af
         assert isinstance(af["data"], Table)
         assert all(af["data"] == table)
 
     # Now test using the table reader
-    new_t = Table.read(tmpfile)
+    new_t = Table.read(tmp_file)
     assert all(new_t == table)
 
 
-def test_table_io_custom_key(tmpdir):
-    tmpfile = str(tmpdir.join("table.asdf"))
+def test_table_io_custom_key(tmp_path):
+    tmp_file = tmp_path / "table.asdf"
 
     table = make_table()
 
-    table.write(tmpfile, data_key="something")
+    table.write(tmp_file, data_key="something")
 
     # Simple sanity check using ASDF directly
-    with asdf.open(tmpfile) as af:
+    with asdf.open(tmp_file) as af:
         assert "something" in af
         assert "data" not in af.keys()
         assert isinstance(af["something"], Table)
@@ -48,24 +48,24 @@ def test_table_io_custom_key(tmpdir):
 
     # Now test using the table reader
     with pytest.raises(KeyError):
-        new_t = Table.read(tmpfile)
+        new_t = Table.read(tmp_file)
 
-    new_t = Table.read(tmpfile, data_key="something")
+    new_t = Table.read(tmp_file, data_key="something")
     assert all(new_t == table)
 
 
-def test_table_io_custom_tree(tmpdir):
-    tmpfile = str(tmpdir.join("table.asdf"))
+def test_table_io_custom_tree(tmp_path):
+    tmp_file = tmp_path / "table.asdf"
 
     table = make_table()
 
     def make_custom_tree(tab):
         return {"foo": {"bar": tab}}
 
-    table.write(tmpfile, make_tree=make_custom_tree)
+    table.write(tmp_file, make_tree=make_custom_tree)
 
     # Simple sanity check using ASDF directly
-    with asdf.open(tmpfile) as af:
+    with asdf.open(tmp_file) as af:
         assert "foo" in af
         assert "bar" in af["foo"]
         assert "data" not in af.keys()
@@ -73,12 +73,12 @@ def test_table_io_custom_tree(tmpdir):
 
     # Now test using table reader
     with pytest.raises(KeyError):
-        new_t = Table.read(tmpfile)
+        new_t = Table.read(tmp_file)
 
     def find_table(asdffile):
         return asdffile["foo"]["bar"]
 
-    new_t = Table.read(tmpfile, find_table=find_table)
+    new_t = Table.read(tmp_file, find_table=find_table)
     assert all(new_t == table)